Split value graph core modules
